Improve extra intuitive ring trigger bindings
This is hard to explain, but in some cases when the player opens a ring, they will be holding some amalgamation of modifiers, and then they might release one or two of them (or keep holding them) and the mounted trigger bindings do not fire because they are combined with modifiers. This should help with the "intuition" of using the rings.