Remove feature to disable modules when disabled in addon list
This is a good idea on paper, to tie the feature of the module into the addon list, but it failed to take an important thing into account; players will often disable all modules belonging to an addon individually when trying to turn the main module off. This lead to many complaints about the interface cursor not working because they wanted to toggle the main addon off to play kb/m, but ticked off every module in the list to do so.