Added handling for macro mismatch in utility ring stored data
Rings use macro ID to execute macros, but these IDs drift as macros are added or removed, which will change their content. We need to keep using macro IDs, but in order to keep IDs up to date, ring validation has been added which tries to match macro content to new IDs when they drift. This is determined by the macro name, icon and body.