Add new method to clear radial focus after using a menu
The new method is based on a deadzone. By fully releasing the stick used to point at a radial object, the deadzone can signal that propagation can continue immediately instead of waiting for a fixed amount of time. With this change, the default time delay has been raised to 1.5 seconds.